app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
bcrypt==4.0.1 
boto3==1.37.21
numpy
scikit-learn
pdfplumber
sentence_transformers
langchain 
langchain-google-genai 
pypdf
langchain-community
langchain-groq
docx2txt
python-decouple
PyPDF2
spacy
langchain-core
groq